10502: extract dbreader/writer from establishing a connection
